app/requirements.txt
Flask
tensorflow-cpu
Pillow
gunicorn
six
typing-extensions
jinja2
numpy
